refactor: organize demos and rename main QADI tool
- Renamed qadi_google_fast.py to qadi.py as the main tool
- Moved all demo and test files to demos/ folder
- Created demos/test_files/ for test scripts
- Created demos/qadi_experiments/ for experimental versions
- Added QADI_README.md with usage instructions
- Cleaned up root directory to keep only essential files

The main QADI tool is now simply:
  uv run python qadi.py "Your question here"
