--- GOOGLE_API_ANALYSIS.md
@@ -1,111 +1,0 @@
-# Google API Integration Analysis
-
-## Summary
-
-The Google API integration is **working correctly**. The timeout issues are caused by performance characteristics, not integration problems.
-
-## Root Cause Analysis
-
-### 1. API Response Times
-- **Google API**: ~4-7 seconds per call
-- **OpenAI/Anthropic**: ~1-2 seconds per call
-- Google's Gemini models are significantly slower than competitors
-
-### 2. Multiple LLM Calls per Phase
-Each QADI phase (Questioning, Abduction, Deduction, Induction) makes multiple LLM calls:
-- Domain analysis (1 call)
-- Strategy-based generation (2-3 calls)
-- Ranking and selection (1 call)
-- **Total**: ~5 LLM calls per phase
-
-### 3. Parallel Execution
-- The system runs all 4 QADI phases in parallel by default
-- This creates 20 concurrent LLM calls (4 phases × 5 calls each)
-- With Google's slow response time, this can take 35+ seconds just for API calls
-
-### 4. Timeout Configuration
-- Default timeouts are 90-120 seconds
-- While technically sufficient, any network delays or rate limiting pushes it over the edge
-
-## Solutions
-
-### 1. Sequential Execution (Recommended for Google)
-```python
-# Instead of parallel execution, run phases one at a time
-for method in [ThinkingMethod.QUESTIONING, ThinkingMethod.ABDUCTION, ...]:
-    result = await orchestrator.run_single_phase(method, ...)
-```
-
-### 2. Reduce LLM Calls
-Configure agents to skip optional analysis:
-```python
-generation_config = {
-    "skip_domain_analysis": True,  # Skip the domain analysis step
-    "max_strategies": 1,           # Use only one strategy instead of 3-4
-    "quick_mode": True            # Simplified generation
-}
-```
-
-### 3. Use Faster Models
-```python
-# Use gemini-1.5-flash instead of gemini-2.5-flash
-os.environ["GEMINI_MODEL_OVERRIDE"] = "gemini-1.5-flash"
-```
-
-### 4. Provider-Aware Configuration
-The system should adapt its behavior based on the LLM provider:
-```python
-if provider == LLMProvider.GOOGLE:
-    config["execution_mode"] = "sequential"
-    config["max_concurrent_requests"] = 2
-else:
-    config["execution_mode"] = "parallel"
-    config["max_concurrent_requests"] = 10
-```
-
-## Performance Comparison
-
-| Provider | Response Time | 4 Phases Parallel | 4 Phases Sequential |
-|----------|--------------|-------------------|---------------------|
-| OpenAI | ~1.5s/call | ~7.5s total | ~30s total |
-| Anthropic | ~2s/call | ~10s total | ~40s total |
-| Google | ~5s/call | ~25s total | ~100s total |
-
-## Recommendations
-
-1. **For Production**: Implement provider-aware orchestration that automatically uses sequential execution for slower providers
-
-2. **For Google Specifically**:
-   - Use sequential execution
-   - Prefer `gemini-1.5-flash` over newer models
-   - Implement caching for domain analysis
-   - Consider hybrid approach: use Google for some phases, faster providers for others
-
-3. **Long-term**: 
-   - Add configuration options for execution mode (parallel/sequential)
-   - Implement adaptive timeout based on provider
-   - Add progress indicators for long-running operations
-   - Consider streaming responses for better UX
-
-## Test Results
-
-```bash
-# Basic Google API Test
-✅ API responds correctly in ~4.7s
-✅ Correct JSON parsing and response handling
-✅ Cost calculation working
-
-# QADI Integration
-❌ Timeouts with default parallel execution
-✅ Works with sequential execution
-✅ Works with reduced LLM calls
-```
-
-## Conclusion
-
-The Google API integration is implemented correctly. The timeouts are due to:
-1. Google's inherently slower API response times
-2. The system making many LLM calls per phase
-3. Parallel execution multiplying the slowness effect
-
-The solution is to adapt the orchestration strategy based on the LLM provider's performance characteristics.
