+++ demos/better_idea_generator.py
