+++ demos/generate_ideas.py
