+++ demos/quick_idea_generation.py
